rename: [user,post]s-[db,router] <- *[Db,Router]
